third_party/blink/web_tests/platform/generic/fast/media/media-query-listener-gc-expected.txt

Media query listeners should work even after gc.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS listenersCalled is expectedResult
PASS successfullyParsed is true

TEST COMPLETE
